bugfix/#5 update readme git submodule command and move args to own package
